services/ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.14.3
google-cloud-vision==3.7.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
redis==5.0.3
psycopg2-binary==2.9.9
pillow==10.2.0
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp==3.9.3
sentry-sdk==1.44.0
